// frontend/src/app/models/tickerInfo.ts
export interface ITickerInfo {
    name: string;
    ticker: string;
    country: string;
    currency: string;
    exchange:string;
    finnhubIndustry: string;
    ipo: string;
    logo: string;
    // marketCapitalization: 
    phone: string;
    // shareOutstanding: 16319.44
    weburl: string;
}